Rename leshan-client-example to leshan-client-demo.
