+++ pom.xml
@@ -48,8 +48,8 @@ Contributors:
 
         <module>leshan-integration-tests</module> 
 
-        <module>leshan-client-example</module>
         <module>leshan-standalone</module>
+        <module>leshan-client-demo</module>
         <module>leshan-bs-server</module>
     </modules>
